refactor agents list
