+++ call-center/web-client/src/components/Agents/styles.css
@@ -1,18 +1,3 @@
-.Agents {
-}
-
-.Agents-heading {
-  font-size: 2rem;
-  font-weight: normal;
-  line-height: 1;
-  margin-top: 0;
-  margin-bottom: 1.5rem;
-}
-
-.Agents-error {
-  text-color: red;
-}
-
 .Agents-list {
   margin: 0;
   padding: 0;
